% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-checks.R
\name{all_pkg_fns_have_exs}
\alias{all_pkg_fns_have_exs}
\title{Check whether all functions have examples}
\usage{
all_pkg_fns_have_exs(path)
}
\arguments{
\item{path}{Path to local source repository}
}
\value{
Vector of named logical values, one for each '.Rd' file indicating
whether or not it has example lines.
}
\description{
Check whether all functions have examples
}
