% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{check_cache}
\alias{check_cache}
\title{check_cache}
\usage{
check_cache(org, repo, cache_dir = tempdir())
}
\arguments{
\item{org}{Github organization}

\item{repo}{Github repository}

\item{cache_dir}{Directory in which packages are cached}
}
\value{
FALSE If a package has previously been cached, and the github repo
has not changed; otherwise TRUE.
}
\description{
Check whether a package has been cached, and if so, whether commits have been
added to the github repo since cached version.
}
\note{
This function is not intended to be called directly, and is only
exported to enable it to be used within the \pkg{plumber} API.
}
