% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-checks.R
\name{ci_badges}
\alias{ci_badges}
\title{Get all CI badges from a repository}
\usage{
ci_badges(u)
}
\arguments{
\item{u}{URL of repo}
}
\value{
Character vector of hyperlinked badge images
}
\description{
Get all CI badges from a repository
}
