% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-checks.R
\name{ci_results_gh}
\alias{ci_results_gh}
\title{CI results for GitHub only}
\usage{
ci_results_gh(path)
}
\arguments{
\item{path}{Path to local source repository}
}
\value{
A 'data.frame' with one row for each GitHub workflow, and columns for
name, the 'conclusion' status, the git 'sha', and the date.
}
\description{
CI results for GitHub only
}
