% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_repo.R
\name{dl_gh_repo}
\alias{dl_gh_repo}
\title{Download a GitHub repo to local cache}
\usage{
dl_gh_repo(u)
}
\arguments{
\item{u}{URL of GitHub repository}
}
\value{
Path to locally cached '.zip' version of repository
}
\description{
Download a GitHub repo to local cache
}
