% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{get_default_branch}
\alias{get_default_branch}
\title{get_default_branch}
\usage{
get_default_branch(org, repo)
}
\arguments{
\item{org}{Github organization}

\item{repo}{Github repository}
}
\value{
Name of default branch on GitHub
}
\description{
get_default_branch
}
\note{
This function is not intended to be called directly, and is only
exported to enable it to be used within the \pkg{plumber} API.
}
\seealso{
Other github: 
\code{\link{get_gh_token}()},
\code{\link{get_latest_commit}()}
}
\concept{github}
