% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodpractice.R
\name{get_gp_report}
\alias{get_gp_report}
\title{Get 'goodpractice' report on local source package}
\usage{
get_gp_report(local_repo)
}
\arguments{
\item{local_repo}{Path to local source of repository}
}
\value{
A \pkg{goodpractice} report
}
\description{
This uses a caching system to only generate new report if repository has been
updated, otherwise it returns locally cached version.
}
