% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{get_latest_commit}
\alias{get_latest_commit}
\title{get_latest_commit}
\usage{
get_latest_commit(org, repo, branch = NULL)
}
\arguments{
\item{org}{Github organization}

\item{repo}{Github repository}

\item{branch}{Branch from which to get latest commit}
}
\value{
Details of latest commit including OID hash
}
\description{
get_latest_commit
}
\note{
This returns the latest commit from the default branch as specified on
GitHub, which will not necessarily be the same as information returned from
\code{gert::git_info} if the \code{HEAD} of a local repository does not point to the
same default branch.
}
\examples{
\dontrun{
org <- "ropensci-review-tools"
repo <- "pkgcheck"
commit <- get_latest_commit (org, repo)
}
}
\seealso{
Other github:
\code{\link[=get_default_github_branch]{get_default_github_branch()}},
\code{\link[=use_github_action_pkgcheck]{use_github_action_pkgcheck()}}
}
\concept{github}
