% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-checks.R
\name{left_assign}
\alias{left_assign}
\title{Check that left-assignment operators are used consistently throughout a
package. "LEFT_ASSIGN" tokens can also be \verb{:=}, so these must also be
tallied, but are ignored.
Left-assign operators are: \link{"=", "<-", "<<-", ":="}.
https://github.com/wch/r-source/blob/trunk/src/main/gram.y#L3283-L3290
https://github.com/wch/r-source/blob/trunk/src/main/gram.y#L3346-L3349}
\usage{
left_assign(path)
}
\arguments{
\item{path}{Path to local source repository}
}
\value{
Named vector of 2 values tallying instances of usage of \verb{<-} and \code{=}.
}
\description{
Check that left-assignment operators are used consistently throughout a
package. "LEFT_ASSIGN" tokens can also be \verb{:=}, so these must also be
tallied, but are ignored.
Left-assign operators are: \link{"=", "<-", "<<-", ":="}.
https://github.com/wch/r-source/blob/trunk/src/main/gram.y#L3283-L3290
https://github.com/wch/r-source/blob/trunk/src/main/gram.y#L3346-L3349
}
