% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_lines.R
\name{num_ex_lines}
\alias{num_ex_lines}
\title{Count number of example lines per function}
\usage{
num_ex_lines(local_repo)
}
\arguments{
\item{local_repo}{Path to local source repository}
}
\value{
A \code{data.frame} of function names and counts of numbers of lines of
example code.
}
\description{
Count number of example lines per function
}
