% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-checks.R
\name{pkg_has_contrib_md}
\alias{pkg_has_contrib_md}
\title{Check whether package has 'contributing.md', as well as whether it has a
life cycle statement}
\usage{
pkg_has_contrib_md(path)
}
\arguments{
\item{path}{Path to local source repository}
}
\value{
Vector of two logical values, the first one identifying whether a
package has a 'contributing.md' file or not; the second whether this file has
a life cycle statement
}
\description{
Check whether package has 'contributing.md', as well as whether it has a
life cycle statement
}
