% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-checks.R
\name{pkg_uses_roxygen2}
\alias{pkg_uses_roxygen2}
\title{Check whether all man files have been generated by 'roxygen2'}
\usage{
pkg_uses_roxygen2(path)
}
\arguments{
\item{path}{Path to local source repository}
}
\value{
'TRUE' if all files generated by 'roxygen2', otherwise 'FALSE'
}
\description{
Check whether all man files have been generated by 'roxygen2'
}
