% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgcheck-bg.R
\name{pkgcheck_bg}
\alias{pkgcheck_bg}
\title{Generate report on package compliance with rOpenSci Statistical Software
requirements as background process}
\usage{
pkgcheck_bg(path)
}
\arguments{
\item{path}{Path to local repository}
}
\value{
A \pkg{processx} object connecting to the background process
generating the main \link{pkgcheck} results (see Note).
}
\description{
Generate report on package compliance with rOpenSci Statistical Software
requirements as background process
}
\note{
The return object will by default display whether it is still running,
or whether it has finished. Once it has finished, the results can be obtained
by calling \verb{$get_result()}, or the main \link{pkgcheck} function can be
called to quickly retrieve the main results from local cache.

This function does not accept the \code{extra_env} parameter of the main
\link{pkgcheck} function, and can not be used to run extra, locally-defined
checks.
}
\examples{
f <- system.file ("extdata", "pkgstats_9.9.tar.gz", package = "pkgstats")
path <- pkgstats::extract_tarball (f)
on.exit (fs::dir_delete (path))

\donttest{
# Foreground checks as "blocking" process which will return
# only after all checks have finished:
checks <- pkgcheck (path, goodpractice = FALSE)
}

# Or run process in background, do other things in the meantime,
# and obtain checks once they have finished:
\dontrun{
ps <- pkgcheck_bg (path)
ps # print status to screen, same as 'ps$print()'
# Once finished, 'pkgcheck' results can be extracted with:
checks <- ps$get_result ()
}
}
\seealso{
Other pkgcheck_fns:
\code{\link[=pkgcheck]{pkgcheck()}},
\code{\link[=print.pkgcheck]{print.pkgcheck()}}
}
\concept{pkgcheck_fns}
