% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-data.R
\name{pkgcheck_update_pkgstats_data}
\alias{pkgcheck_update_pkgstats_data}
\title{rOpenSci-internal function to update \pkg{pkgstats} data.}
\usage{
pkgcheck_update_pkgstats_data()
}
\description{
Calling this function will create an updated version of this repository in
the R temporary directory, and have no other effect for unauthorized users.
}
