% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{pkgrep_install_deps}
\alias{pkgrep_install_deps}
\title{Install all system and package dependencies of an R package}
\usage{
pkgrep_install_deps(local_repo, os, os_release)
}
\arguments{
\item{local_repo}{Path to local package}

\item{os}{Name of operating system, passed to \pkg{remotes} package to
install system dependencies.}

\item{os_release}{Name of operating system release, passed to \pkg{remotes}
package to install system dependencies.}
}
\description{
Install all system and package dependencies of an R package
}
