% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srr.R
\name{pkgrep_srr_report}
\alias{pkgrep_srr_report}
\title{Format the \pkg{srr} reporting section of the general editorial report.}
\usage{
pkgrep_srr_report(local_repo)
}
\arguments{
\item{local_repo}{Location of local repository to report on}
}
\description{
Format the \pkg{srr} reporting section of the general editorial report.
}
