% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgreport-package.R
\docType{package}
\name{pkgreport-package}
\alias{pkgreport}
\alias{pkgreport-package}
\title{pkgreport: Plumber API to report package structure and function}
\description{
Plumber API to report package structure and
    function.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropenscilabs/pkgreport}
  \item Report bugs at \url{https://github.com/ropenscilabs/pkgreport/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

}
\keyword{internal}
