% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgreport-fn.R
\name{pkgreport}
\alias{pkgreport}
\title{Generate report on package compliance with rOpenSci Statistical Software
requirements}
\usage{
pkgreport(path)
}
\arguments{
\item{path}{Path to local repository}
}
\value{
A \code{pkgreport} object detailing all package assessments automatically
applied to packages submitted for peer review.
}
\description{
Generate report on package compliance with rOpenSci Statistical Software
requirements
}
