% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgreport-fn.R
\name{pkgstats_checks}
\alias{pkgstats_checks}
\title{Format \pkg{pkgstats} data}
\usage{
pkgstats_checks(s)
}
\arguments{
\item{s}{Output of \pkg{pkgstats} call.}
}
\value{
Report as formatted string
}
\description{
Format \pkg{pkgstats} data
}
