% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgreport-package.R
\docType{data}
\name{pkgstats_data}
\alias{pkgstats_data}
\title{Statistics on all CRAN packages from 'pkgstats'}
\format{
An object of class \code{data.frame} with 14834 rows and 81 columns.
}
\usage{
pkgstats_data
}
\description{
Statistics on all CRAN packages from 'pkgstats'
}
\keyword{datasets}
