% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{pr_dl}
\alias{pr_dl}
\title{pr_dl}
\usage{
pr_dl(u, port = 8000)
}
\arguments{
\item{u}{URL of package repository}

\item{port}{Port at which plumber API has been served via \link{serve_api}.}
}
\value{
Location of download, to be passed to \link{pg_report}
}
\description{
Download a package from a specified URL, storing in tempdir()
}
