% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{pr_report}
\alias{pr_report}
\title{pr_report}
\usage{
pr_report(dl, port = 8000)
}
\arguments{
\item{dl}{Location of downloaded repository as returned from \link{pr_dl}}

\item{port}{Port used to set up API with \link{serve_api}}
}
\description{
Query plumber API (set up with \link{serve_api}) for a report on a repository
downloaded locally with the \link{pr_dl} function.
}
