% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-books.R
\name{read_pkg_guide}
\alias{read_pkg_guide}
\title{Browse packaging guidelines}
\usage{
read_pkg_guide(which = c("release", "dev"))
}
\arguments{
\item{which}{Whether to read the released or "dev" development version.}
}
\value{
Nothing. Function called purely for side-effect of opening web page
with package guidelines.
}
\description{
A convenience function to automatically open the web page of rOpenSci's
"Package Development Guide" in the default browser.
}
\examples{
\donttest{
read_pkg_guide ()
}
}
