% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-checks.R
\name{render_markdown}
\alias{render_markdown}
\title{render markdown-formatted input into 'html'}
\usage{
render_markdown(md, open = TRUE)
}
\arguments{
\item{md}{Result of \link{checks_to_markdown} function.}

\item{open}{If \code{TRUE}, open \code{hmtl}-rendered version in web browser.}
}
\value{
(invisible) Location of \code{.html}-formatted version of input.
}
\description{
render markdown-formatted input into 'html'
}
\seealso{
Other extra: 
\code{\link{checks_to_markdown}()},
\code{\link{logfile_names}()}
}
\concept{extra}
