% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-checks.R
\name{stats_checks}
\alias{stats_checks}
\title{Compare 'pkgstats' summary with statistics from all CRAN packages.}
\usage{
stats_checks(s, threshold = 0.05)
}
\arguments{
\item{s}{Result of \link{pkgstats_summary}}

\item{threshold}{Proportion threshold below which to report on statistically
unusual properties.}
}
\value{
A 'data.frame' of selected statistical properties and percentiles in
relation to all other packages on CRAN.
}
\description{
Compare 'pkgstats' summary with statistics from all CRAN packages.
}
