% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-checks.R
\name{checks_to_markdown}
\alias{checks_to_markdown}
\title{Convert checks to markdown-formatted report}
\usage{
checks_to_markdown(checks, render = FALSE)
}
\arguments{
\item{checks}{Result of main \link{pkgcheck} function}

\item{render}{If \code{TRUE}, render output as \code{html} document and open in
browser.}
}
\value{
Markdown-formatted version of check report
}
\description{
Convert checks to markdown-formatted report
}
\examples{
\dontrun{
checks <- pkgcheck ("/path/to/my/package")
md <- checks_to_markdown (checks) # markdown-formatted character vector
md <- checks_to_markdown (checks, render = TRUE) # HTML version
}
}
\seealso{
Other extra: 
\code{\link{fn_names_on_cran}()},
\code{\link{list_pkgchecks}()},
\code{\link{logfile_names}()},
\code{\link{render_md2html}()}
}
\concept{extra}
