% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodpractice.R
\name{convert_gp_components}
\alias{convert_gp_components}
\title{Convert \pkg{goodpractice} components into templated report}
\usage{
convert_gp_components(x)
}
\arguments{
\item{x}{List of components of \pkg{goodpractice} report}
}
\value{
Markdown-formatted report of contents of \code{x}
}
\description{
Convert \pkg{goodpractice} components into templated report
}
