% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor-check.R
\name{editor_check}
\alias{editor_check}
\title{Body of main 'editorcheck' response}
\usage{
editor_check(path, u)
}
\arguments{
\item{path}{Path to local source repository}

\item{u}{URL of package repository}
}
\value{
Markdown-formatted response body from static package checks.
}
\description{
Body of main 'editorcheck' response
}
