% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-unique-fn-names.R
\name{fn_names_on_cran}
\alias{fn_names_on_cran}
\title{Check whether a function name exists in any CRAN packages}
\usage{
fn_names_on_cran(fn_name, force_update = FALSE)
}
\arguments{
\item{fn_name}{Character vector of one or more function names to check.}

\item{force_update}{If 'TRUE', locally-cached data of all function names
from all CRAN packages will be updated to latest version.}
}
\value{
A \code{data.frame} of three columns, "package", "version", and
"fn_name", identifying any other packages matching specified function
name(s). If no matches are found, the \code{data.frame} will have no rows.
}
\description{
Check whether a function name exists in any CRAN packages
}
\examples{
fn_names_on_cran (c ("min", "max"))

}
\seealso{
Other extra: 
\code{\link{checks_to_markdown}()},
\code{\link{list_pkgchecks}()},
\code{\link{logfile_names}()},
\code{\link{render_md2html}()}
}
\concept{extra}
