% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{get_default_github_branch}
\alias{get_default_github_branch}
\title{get_default_github_branch}
\usage{
get_default_github_branch(org, repo)
}
\arguments{
\item{org}{Github organization}

\item{repo}{Github repository}
}
\value{
Name of default branch on GitHub
}
\description{
get_default_github_branch
}
\note{
This function is not intended to be called directly, and is only
exported to enable it to be used within the \pkg{plumber} API.
}
\examples{
\dontrun{
org <- "ropensci-review-tools"
repo <- "pkgcheck"
branch <- get_default_github_branch (org, repo)
}
}
\seealso{
Other github:
\code{\link[=get_latest_commit]{get_latest_commit()}},
\code{\link[=use_github_action_pkgcheck]{use_github_action_pkgcheck()}}
}
\concept{github}
