% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{get_gh_token}
\alias{get_gh_token}
\title{Get GitHub token}
\usage{
get_gh_token(token_name = "")
}
\arguments{
\item{token_name}{Optional name of token to use}
}
\value{
The value of the GitHub access token extracted from environment
variables.
}
\description{
Get GitHub token
}
