% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{get_gh_token}
\alias{get_gh_token}
\title{Get GitHub token}
\usage{
get_gh_token(token_name = "")
}
\arguments{
\item{token_name}{Optional name of token to use}
}
\value{
The value of the GitHub access token extracted from environment
variables.
}
\description{
Get GitHub token
}
\seealso{
Other github: 
\code{\link{get_default_branch}()},
\code{\link{get_latest_commit}()},
\code{\link{use_github_action_pkgcheck}()}
}
\concept{github}
