% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{get_latest_commit}
\alias{get_latest_commit}
\title{get_latest_commit}
\usage{
get_latest_commit(org, repo)
}
\arguments{
\item{org}{Github organization}

\item{repo}{Github repository}
}
\value{
Details of latest commit including OID hash
}
\description{
get_latest_commit
}
\note{
This function is not intended to be called directly, and is only
exported to enable it to be used within the \pkg{plumber} API.
}
