% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_pkgchecks}
\alias{list_pkgchecks}
\title{Print to screen a list of all checks currently implemented}
\usage{
list_pkgchecks()
}
\value{
Character vector of names of all checks (invisibly)
}
\description{
Print to screen a list of all checks currently implemented
}
