% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_pkgchecks}
\alias{list_pkgchecks}
\title{List all checks currently implemented}
\usage{
list_pkgchecks(quiet = FALSE)
}
\arguments{
\item{quiet}{If \code{TRUE}, print all checks to screen. Function invisibly
returns list of checks regardless.}
}
\value{
Character vector of names of all checks (invisibly)
}
\description{
List all checks currently implemented
}
\examples{
list_pkgchecks ()
}
\seealso{
Other extra: 
\code{\link{checks_to_markdown}()},
\code{\link{logfile_names}()},
\code{\link{render_md2html}()}
}
\concept{extra}
