% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{logfile_names}
\alias{logfile_names}
\title{Set up stdout & stderr cache files for \code{r_bg} process}
\usage{
logfile_names(path)
}
\arguments{
\item{path}{Path to local repository}
}
\value{
Vector of two strings holding respective local paths to \code{stdout} and
\code{stderr} files for \code{r_bg} process controlling the main \link{pkgcheck}
function when executed in background mode.
}
\description{
Set up stdout & stderr cache files for \code{r_bg} process
}
\note{
These files are needed for the \pkg{callr} \code{r_bg} process which
controls the main \link{pkgcheck}. The \code{stdout} and \code{stderr} pipes from the
process are stored in the cache directory so they can be inspected via their
own distinct endpoint calls.
}
\examples{
\dontrun{
logfiles <- logfiles_namnes ("/path/to/my/package")
print (logfiles)
}
}
\seealso{
Other extra:
\code{\link[=checks_to_markdown]{checks_to_markdown()}},
\code{\link[=fn_names_on_cran]{fn_names_on_cran()}},
\code{\link[=list_pkgchecks]{list_pkgchecks()}},
\code{\link[=render_md2html]{render_md2html()}}
}
\concept{extra}
