% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-checks.R
\name{pkg_desc_has_url}
\alias{pkg_desc_has_url}
\title{Check whether package 'DESCRIPTION' file has URL + BugReports fields}
\usage{
pkg_desc_has_url(path)
}
\arguments{
\item{path}{Path to local source repository}
}
\description{
Check whether package 'DESCRIPTION' file has URL + BugReports fields
}
