% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-checks.R
\name{pkg_has_contrib_md}
\alias{pkg_has_contrib_md}
\title{Check whether package has 'contributing.md', as well as whether it has a
life cycle statement}
\usage{
pkg_has_contrib_md(path, check_lifecycle = TRUE)
}
\arguments{
\item{path}{Path to local source repository}
}
\description{
Check whether package has 'contributing.md', as well as whether it has a
life cycle statement
}
