% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgcheck-package.R
\docType{package}
\name{pkgcheck-package}
\alias{pkgcheck-package}
\title{pkgcheck: Package Checks for 'rOpenSci'}
\description{
Check whether a package is ready for submission to the 'rOpenSci' peer review system ('rOpenSci' authors, 2021; \doi{10.5281/zenodo.5556756}). Incorporates the 'goodpractice' package and many additional checks, including aspects related to maintenance of online public code repositories.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/pkgcheck/}
  \item \url{https://github.com/ropensci-review-tools/pkgcheck}
  \item Report bugs at \url{https://github.com/ropensci-review-tools/pkgcheck/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com} (\href{https://orcid.org/0000-0003-2172-5265}{ORCID})

Authors:
\itemize{
  \item Mark Padgham \email{mark.padgham@email.com} (\href{https://orcid.org/0000-0003-2172-5265}{ORCID})
  \item Maëlle Salmon
  \item Jacob Wujciak-Jens \email{jacob@wujciak.de} (\href{https://orcid.org/0000-0002-7281-3989}{ORCID})
  \item Eunseop Kim \email{markean@pm.me} (\href{https://orcid.org/0009-0000-2138-788X}{ORCID})
  \item Andy Teucher \email{andy.teucher@gmail.com} (\href{https://orcid.org/0000-0002-7840-692X}{ORCID})
}

Other contributors:
\itemize{
  \item Kelli F. Johnson \email{kelli.johnson@noaa.gov} (\href{https://orcid.org/0000-0002-5149-451X}{ORCID}) [contributor]
  \item Katrina Brock \email{katrinabrock266@gmail.com} (\href{https://orcid.org/0009-0003-4678-9545}{ORCID}) [contributor]
}

}
\keyword{internal}
