% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgcheck-package.R
\docType{package}
\name{pkgcheck-package}
\alias{pkgcheck-package}
\alias{_PACKAGE}
\title{pkgcheck: rOpenSci Package Checks}
\description{
Check whether a package is ready for submission to rOpenSci’s
    peer review system.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci-review-tools/pkgcheck}
  \item Report bugs at \url{https://github.com/ropensci-review-tools/pkgcheck/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

}
\keyword{internal}
