% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgcheck-package.R
\docType{package}
\name{pkgcheck-package}
\alias{pkgcheck-package}
\title{pkgcheck: rOpenSci Package Checks}
\description{
Check whether a package is ready for submission to rOpenSci's peer review system.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/pkgcheck/}
  \item \url{https://github.com/ropensci-review-tools/pkgcheck}
  \item Report bugs at \url{https://github.com/ropensci-review-tools/pkgcheck/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com} (\href{https://orcid.org/0000-0003-2172-5265}{ORCID})

Authors:
\itemize{
  \item Maëlle Salmon
  \item Jacob Wujciak-Jens \email{jacob@wujciak.de} (\href{https://orcid.org/0000-0002-7281-3989}{ORCID})
  \item Eunseop Kim \email{markean@pm.me} (\href{https://orcid.org/0009-0000-2138-788X}{ORCID})
}

}
\keyword{internal}
