% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgcheck-fn.R
\name{pkgcheck}
\alias{pkgcheck}
\title{Generate report on package compliance with rOpenSci Statistical Software
requirements}
\usage{
pkgcheck(
  path = ".",
  goodpractice = TRUE,
  use_cache = TRUE,
  extra_env = .GlobalEnv
)
}
\arguments{
\item{path}{Path to local repository}

\item{goodpractice}{If \code{FALSE}, skip goodpractice checks. May be useful in
development stages to more quickly check other aspects.}

\item{use_cache}{Checks are cached for rapid retrieval, and only re-run if
the git hash of the local repository changes. Setting \code{use_cache} to \code{FALSE}
will for checks to be re-run even if the git hash has not changed.}

\item{extra_env}{Additional environments from which to collate checks. Other
package names may be appended using \code{c}, as in \code{c(.GlobalEnv, "mypkg")}.}
}
\value{
A \code{pkgcheck} object detailing all package assessments automatically
applied to packages submitted for peer review.
}
\description{
Generate report on package compliance with rOpenSci Statistical Software
requirements
}
\seealso{
Other pkgcheck_fns: 
\code{\link{pkgcheck_bg}()}
}
\concept{pkgcheck_fns}
