% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgcheck-fn.R
\name{pkgcheck}
\alias{pkgcheck}
\title{Generate report on package compliance with rOpenSci Statistical Software
requirements}
\usage{
pkgcheck(path = ".")
}
\arguments{
\item{path}{Path to local repository}
}
\value{
A \code{pkgcheck} object detailing all package assessments automatically
applied to packages submitted for peer review.
}
\description{
Generate report on package compliance with rOpenSci Statistical Software
requirements
}
\seealso{
Other pkgcheck_fns: 
\code{\link{pkgcheck_bg}()}
}
\concept{pkgcheck_fns}
