% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgcheck-fn.R
\name{pkgcheck}
\alias{pkgcheck}
\title{Generate report on package compliance with rOpenSci Statistical Software
requirements}
\usage{
pkgcheck(path = ".", goodpractice = TRUE, extra_env = .GlobalEnv)
}
\arguments{
\item{path}{Path to local repository}

\item{goodpractice}{If \code{FALSE}, skip goodpractice checks. May be useful in
development stages to more quickly check other aspects.}

\item{extra_env}{Additional environments from which to collate checks. Other
package names may be appended using \code{c}, as in \code{c(.GlobalEnv, "mypkg")}.}
}
\value{
A \code{pkgcheck} object detailing all package assessments automatically
applied to packages submitted for peer review.
}
\description{
Generate report on package compliance with rOpenSci Statistical Software
requirements
}
\seealso{
Other pkgcheck_fns: 
\code{\link{pkgcheck_bg}()}
}
\concept{pkgcheck_fns}
