% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgcheck-bg.R
\name{pkgcheck_bg}
\alias{pkgcheck_bg}
\title{Generate report on package compliance with rOpenSci Statistical Software
requirements as background process}
\usage{
pkgcheck_bg(path)
}
\arguments{
\item{path}{Path to local repository}
}
\value{
A \pkg{processx} object connecting to the background process
generating the main \link{pkgcheck} results (see Note).
}
\description{
Generate report on package compliance with rOpenSci Statistical Software
requirements as background process
}
\note{
The return object will by default display whether it is still running,
or whether it has finished. Once it has finished, the results can be obtained
by calling \verb{$get_result()}, or the main \link{pkgcheck} function can be
called to quickly retrieve the main results from local cache.
}
\seealso{
Other pkgcheck_fns: 
\code{\link{pkgcheck}()}
}
\concept{pkgcheck_fns}
