% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgcheck-package.R
\docType{data}
\name{pkgstats_data}
\alias{pkgstats_data}
\title{Statistics on all CRAN packages from 'pkgstats'}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 21091 rows and 91 columns.
}
\usage{
pkgstats_data
}
\description{
Statistics on all CRAN packages from 'pkgstats'
}
\concept{data}
\keyword{datasets}
