% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgcheck-methods.R
\name{print.pkgcheck}
\alias{print.pkgcheck}
\title{Generic print method for 'pkgcheck' objects.}
\usage{
\method{print}{pkgcheck}(x, deps = FALSE, ...)
}
\arguments{
\item{x}{A 'pkgcheck' object to be printed.}

\item{deps}{If 'TRUE', include details of dependency packages and function
usage.}

\item{...}{Further arguments pass to or from other methods (not used here).}
}
\value{
Nothing. Method called purely for side-effect of printing to screen.
}
\description{
Generic print method for 'pkgcheck' objects.
}
\examples{
f <- system.file ("extdata", "pkgstats_9.9.tar.gz", package = "pkgstats")
path <- pkgstats::extract_tarball (f)

\dontrun{
checks <- pkgcheck (path)
summary (checks) # print summary only
print (checks) # print full checks, starting with summary
}
fs::dir_delete (path)
}
\seealso{
Other pkgcheck_fns:
\code{\link[=pkgcheck]{pkgcheck()}},
\code{\link[=pkgcheck_bg]{pkgcheck_bg()}}
}
\concept{pkgcheck_fns}
