% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgcheck-methods.R
\name{print.pkgcheck}
\alias{print.pkgcheck}
\title{Generic print method for 'pkgcheck' objects.}
\usage{
\method{print}{pkgcheck}(x, deps = FALSE, ...)
}
\arguments{
\item{x}{A 'pkgcheck' object to be printed.}

\item{deps}{If 'TRUE', include details of dependency packages and function
usage.}

\item{...}{Further arguments pass to or from other methods (not used here).}
}
\description{
Generic print method for 'pkgcheck' objects.
}
