% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodpractice.R
\name{process_gp}
\alias{process_gp}
\title{process 'goodpractice' report}
\usage{
process_gp(gp)
}
\arguments{
\item{gp}{A 'goodpractice' object as returned by \pkg{goodpractice}}
}
\value{
Markdown-formatted report
}
\description{
Extract various components from 'goodpractice' report and convert into format
to be returned by endpoint
}
