% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-books.R
\name{read_pkg_guide}
\alias{read_pkg_guide}
\title{Browse packaging guidelines}
\usage{
read_pkg_guide(which = c("release", "dev"))
}
\arguments{
\item{which}{Whether to read the released or "dev" development version.}
}
\description{
Browse packaging guidelines
}
