% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-checks.R
\name{render_md2html}
\alias{render_md2html}
\title{render markdown-formatted input into 'html'}
\usage{
render_md2html(md, open = TRUE)
}
\arguments{
\item{md}{Result of \link{checks_to_markdown} function.}

\item{open}{If \code{TRUE}, open \code{hmtl}-rendered version in web browser.}
}
\value{
(invisible) Location of \code{.html}-formatted version of input.
}
\description{
render markdown-formatted input into 'html'
}
\examples{
\dontrun{
checks <- pkgcheck ("/path/to/my/package")
# Generate standard markdown-formatted character vector:
md <- checks_to_markdown (checks)

# Directly generate HTML output:
h <- checks_to_markdown (checks, render = TRUE) # HTML version

# Or convert markdown-formatted version to HTML:
h <- render_md2html (md)
}
}
\seealso{
Other extra: 
\code{\link{checks_to_markdown}()},
\code{\link{list_pkgchecks}()},
\code{\link{logfile_names}()}
}
\concept{extra}
