% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{serve_api}
\alias{serve_api}
\title{serve plumber API to report on packages}
\usage{
serve_api(port = 8000L, cache_dir = tempdir(), bg = TRUE, debug = FALSE)
}
\arguments{
\item{port}{Port for API to be served on}

\item{cache_dir}{Directory where previously downloaded repositories are
cached}

\item{bg}{If \code{FALSE}, run process as a blocking foreground process}

\item{debug}{For background processes: If \code{TRUE}, dump output logs to
\verb{/tmp/out} and \verb{/tmp/err}}
}
\value{
A \code{processx} process which must be actively stopped with \code{ps$kill()}.
}
\description{
The API exposes the two POST points of
\itemize{
\item report Download the given URL to \code{tempdir()}
\item pgstats Analyse the package held at location returned from \code{dlurl} and
return a formatted report
}
}
