% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{serve_api}
\alias{serve_api}
\title{serve plumber API to report on packages}
\usage{
serve_api(port = 8000L, cache_dir = NULL, debug = FALSE)
}
\arguments{
\item{port}{Port for API to be served on}

\item{cache_dir}{Directory where previously downloaded repositories are
cached}

\item{debug}{For background processes: If \code{TRUE}, dump output logs to
\verb{/tmp/out} and \verb{/tmp/err}}
}
\value{
Nothing; calling this starts a blocking process.
}
\description{
The API exposes the single POST points of \code{report} to download software from
the given URL and return a textual analysis of its structure and
functionality.
}
